refactor(engine): drop dead code + clippy --fix mechanical lints
Tier A — dead code / stale suppressions:
- Delete `_unused_paths()` (aircraft_extract bin): a no-op that only
  silenced dead_code on read_record_batches/read_flights, both of which
  are now used across stage_1/stage_2c. Drop its two now-unused imports.
- Remove stale `#[allow(dead_code)]` / `#[allow(unused_imports)]` guarding
  airport_summary_part read + re-export — the Stage 2C v5 reduce phase
  they waited for has landed (airport_summary_reduce.rs consumes them in
  production); fix the "next implementation step" comments that no
  longer hold.
- Drop an unused `#[cfg(test)]` import (AirportTrafficRowView) in
  source-reader — the one real rustc warning the incremental cache hid.

Tier B — `cargo clippy --fix`, MachineApplicable suggestions only:
- redundant same-type casts, map_or(false,..)→is_some_and, a|b|c→a..=c
  match arms, needless `as_deref_mut()` on a last use, manual range
  contains, excessive-precision float literal, added Default impls.

Verified byte-identical: 570 engine unit tests pass; 115-point live
popup parity vs the pre-change build shows 0 drift > 1e-9 (max 2e-14,
the f64 summation-order noise floor).

+++ engine/noise-compute/src/emission/aircraft/segment_filters.rs
@@ -146,10 +146,10 @@ pub fn is_valid_airborne_segment(seg: &AircraftSegment, rasters: &dyn RasterSamp
         return false;
     }
 
-    let sl = seg.start_lat as f64;
-    let sn = seg.start_lon as f64;
-    let el = seg.end_lat as f64;
-    let en = seg.end_lon as f64;
+    let sl = seg.start_lat;
+    let sn = seg.start_lon;
+    let el = seg.end_lat;
+    let en = seg.end_lon;
     let sa = seg.start_alt_m as f64;
     let ea = seg.end_alt_m as f64;
     for frac in [0.25_f64, 0.75] {
